value entity registration with provider

+++ codegen/java-gen/src/main/scala/com/lightbend/akkasls/codegen/java/EntityServiceSourceGenerator.scala
@@ -66,15 +66,27 @@ object EntityServiceSourceGenerator {
       interfaceSource(service, entity, packageName, className).getBytes(Charsets.UTF_8)
     )
 
-    val handleClassName = className + "Handler"
+    val handlerClassName = className + "Handler"
     val handlerSourcePath =
-      generatedSourceDirectory.resolve(packagePath.resolve(handleClassName + ".java"))
+      generatedSourceDirectory.resolve(packagePath.resolve(handlerClassName + ".java"))
 
-    handlerSourcePath.getParent.toFile.mkdirs()
     handlerSource(service, entity, packageName, className).foreach { doc =>
+      handlerSourcePath.getParent.toFile.mkdirs()
       Files.write(
         handlerSourcePath,
         doc.getBytes(Charsets.UTF_8)
+      )
+    }
+
+    val providerClassName = className + "Provider"
+    val providerSourcePath =
+      generatedSourceDirectory.resolve(packagePath.resolve(providerClassName + ".java"))
+
+    providerSource(service, entity, packageName, className).foreach { src =>
+      providerSourcePath.getParent.toFile.mkdirs()
+      Files.write(
+        providerSourcePath,
+        src.getBytes(Charsets.UTF_8)
       )
     }
 
@@ -83,16 +95,21 @@ object EntityServiceSourceGenerator {
       val testClassName = className + "Test"
       val testSourcePath =
         testSourceDirectory.resolve(packagePath.resolve(testClassName + ".java"))
-      val testSourceFiles = if (!testSourcePath.toFile.exists()) {
-        testSourcePath.getParent.toFile.mkdirs()
-        Files.write(
-          testSourcePath,
-          testSource(service, entity, packageName, implClassName, testClassName).getBytes(Charsets.UTF_8)
-        )
-        List(testSourcePath)
-      } else {
-        List.empty
-      }
+      val testSourceFiles =
+        if (!testSourcePath.toFile.exists()) {
+          testSource(service, entity, packageName, implClassName, testClassName) match {
+            case Some(src) =>
+              testSourcePath.getParent.toFile.mkdirs()
+              Files.write(
+                testSourcePath,
+                src.getBytes(Charsets.UTF_8)
+              )
+              List(testSourcePath)
+            case None => List.empty
+          }
+        } else {
+          List.empty
+        }
 
       // ...and then its integration test
       val integrationTestClassName = className + "IntegrationTest"
@@ -298,6 +315,18 @@ object EntityServiceSourceGenerator {
     }
   }
 
+  private[codegen] def providerSource(service: ModelBuilder.EntityService,
+                                      entity: ModelBuilder.Entity,
+                                      packageName: String,
+                                      className: String): Option[String] = {
+    entity match {
+      case eventSourcedEntity: ModelBuilder.EventSourcedEntity =>
+        None
+      case valueEntity: ValueEntity =>
+        Some(ValueEntitySourceGenerator.valueEntityProvider(service, valueEntity, packageName, className))
+    }
+  }
+
   private[codegen] def abstractEventSourcedEntity(
       service: ModelBuilder.EntityService,
       entity: ModelBuilder.EventSourcedEntity,
@@ -380,92 +409,98 @@ object EntityServiceSourceGenerator {
       packageName: String,
       implClassName: String,
       testClassName: String
-  ): String = {
-    val messageTypes =
-      service.commands.flatMap(command => Seq(command.inputType, command.outputType)) ++ (entity match {
-        case _: ModelBuilder.EventSourcedEntity =>
-          Seq.empty
-        case ModelBuilder.ValueEntity(_, _, state) => Seq(state.fqn)
-      })
-
-    val imports = (messageTypes.toSeq
-      .filterNot(_.parent.javaPackage == packageName)
-      .map(typeImport) ++ Seq(
-      "org.junit.Test",
-      "org.mockito.*"
-    ) ++ (entity match {
+  ): Option[String] = {
+    entity match {
+      case _: ValueEntity => None
       case _: EventSourcedEntity =>
-        Seq("com.akkaserverless.javasdk.eventsourcedentity.CommandContext")
-      case _: ValueEntity =>
-        Seq("com.akkaserverless.javasdk.valueentity.CommandContext")
-    })).distinct.sorted
+        val messageTypes =
+          service.commands.flatMap(command => Seq(command.inputType, command.outputType)) ++ (entity match {
+            case _: ModelBuilder.EventSourcedEntity =>
+              Seq.empty
+            case ModelBuilder.ValueEntity(_, _, state) => Seq(state.fqn)
+          })
 
-    pretty(
-      initialisedCodeComment <> line <> line <>
-      "package" <+> packageName <> semi <> line <>
-      line <>
-      ssep(
-        imports.to[immutable.Seq].map(pkg => "import" <+> pkg <> semi),
-        line
-      ) <> line <>
-      line <>
-      "import" <+> "static" <+> "org.junit.Assert.assertThrows" <> semi <> line <>
-      line <>
-      `class`("public", testClassName) {
-        "private" <+> "String" <+> "entityId" <+> equal <+> """"entityId1"""" <> semi <> line <>
-        "private" <+> implClassName <+> "entity" <> semi <> line <>
-        "private" <+> (entity match {
-          case ModelBuilder.ValueEntity(_, _, state) =>
-            "CommandContext" <> angles(qualifiedType(state.fqn))
-          case _ =>
-            "CommandContext"
-        }) <+> "context" <+> equal <+> "Mockito.mock(CommandContext.class)" <> semi <> line <>
-        line <>
-        ssep(
-          service.commands.toSeq
-            .map {
-              command =>
-                "@Test" <> line <>
-                method(
-                  "public",
-                  "void",
-                  lowerFirst(command.fqn.name) + "Test",
-                  List.empty,
-                  emptyDoc
-                ) {
-                  "entity" <+> equal <+> "new" <+> implClassName <> parens(
-                    "entityId"
-                  ) <> semi <> line <>
-                  line <>
-                  "// TODO: write your mock here" <> line <>
-                  "// Mockito.when(context.[...]).thenReturn([...]);" <> line <>
-                  line <>
-                  "// TODO: set fields in command, and update assertions to verify implementation" <> line <>
-                  "//" <+> "assertEquals" <> parens(
-                    "[expected]" <> comma <>
-                    line <> "//" <> indent("entity") <> dot <> lowerFirst(
-                      command.fqn.name
-                    ) <> lparen <>
-                    qualifiedType(
-                      command.inputType
-                    ) <> dot <> "newBuilder().build(), context"
-                  ) <> semi <> line <>
-                  "//" <+> rparen <> semi <>
-                  (entity match {
-                    case _: ModelBuilder.EventSourcedEntity =>
-                      line <>
-                      line <>
-                      "// TODO: if you wish to verify events:" <> line <>
-                      "//" <> indent("Mockito.verify(context).emit(event)") <> semi
-                    case _ => emptyDoc
-                  })
-                }
+        val imports = (messageTypes.toSeq
+          .filterNot(_.parent.javaPackage == packageName)
+          .map(typeImport) ++ Seq(
+          "org.junit.Test",
+          "org.mockito.*"
+        ) ++ (entity match {
+          case _: EventSourcedEntity =>
+            Seq("com.akkaserverless.javasdk.eventsourcedentity.CommandContext")
+          case _: ValueEntity =>
+            Seq("com.akkaserverless.javasdk.valueentity.CommandContext")
+        })).distinct.sorted
+
+        Some(
+          pretty(
+            initialisedCodeComment <> line <> line <>
+            "package" <+> packageName <> semi <> line <>
+            line <>
+            ssep(
+              imports.to[immutable.Seq].map(pkg => "import" <+> pkg <> semi),
+              line
+            ) <> line <>
+            line <>
+            "import" <+> "static" <+> "org.junit.Assert.assertThrows" <> semi <> line <>
+            line <>
+            `class`("public", testClassName) {
+              "private" <+> "String" <+> "entityId" <+> equal <+> """"entityId1"""" <> semi <> line <>
+              "private" <+> implClassName <+> "entity" <> semi <> line <>
+              "private" <+> (entity match {
+                case ModelBuilder.ValueEntity(_, _, state) =>
+                  "CommandContext" <> angles(qualifiedType(state.fqn))
+                case _ =>
+                  "CommandContext"
+              }) <+> "context" <+> equal <+> "Mockito.mock(CommandContext.class)" <> semi <> line <>
+              line <>
+              ssep(
+                service.commands.toSeq
+                  .map {
+                    command =>
+                      "@Test" <> line <>
+                      method(
+                        "public",
+                        "void",
+                        lowerFirst(command.fqn.name) + "Test",
+                        List.empty,
+                        emptyDoc
+                      ) {
+                        "entity" <+> equal <+> "new" <+> implClassName <> parens(
+                          "entityId"
+                        ) <> semi <> line <>
+                        line <>
+                        "// TODO: write your mock here" <> line <>
+                        "// Mockito.when(context.[...]).thenReturn([...]);" <> line <>
+                        line <>
+                        "// TODO: set fields in command, and update assertions to verify implementation" <> line <>
+                        "//" <+> "assertEquals" <> parens(
+                          "[expected]" <> comma <>
+                          line <> "//" <> indent("entity") <> dot <> lowerFirst(
+                            command.fqn.name
+                          ) <> lparen <>
+                          qualifiedType(
+                            command.inputType
+                          ) <> dot <> "newBuilder().build(), context"
+                        ) <> semi <> line <>
+                        "//" <+> rparen <> semi <>
+                        (entity match {
+                          case _: ModelBuilder.EventSourcedEntity =>
+                            line <>
+                            line <>
+                            "// TODO: if you wish to verify events:" <> line <>
+                            "//" <> indent("Mockito.verify(context).emit(event)") <> semi
+                          case _ => emptyDoc
+                        })
+                      }
+                  }
+                  .to[immutable.Seq],
+                line <> line
+              )
             }
-            .to[immutable.Seq],
-          line <> line
+          ).layout
         )
-      }
-    ).layout
+    }
   }
 
   private[codegen] def integrationTestSource(
